% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_template.R
\name{yaml_template}
\alias{yaml_template}
\title{Create a YAML template for time logging}
\usage{
yaml_template(copyToClipboard = FALSE)
}
\arguments{
\item{copyToClipboard}{Instead of returning the template text, copy the
template text directly to the clipboard.}
}
\value{
A YAML template to fill out and use with \code{\link{parse_times}}.
}
\description{
Create a YAML template for time logging
}
\examples{
template_text <- yaml_template()
\dontrun{
yaml_template(copyToClipboard = TRUE)
}
}
