% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_times.R
\name{parse_times}
\alias{parse_times}
\title{Parse times from a template of the structure generated by
\code{\link{yaml_template}}}
\usage{
parse_times(times, returnIntermediateTable = FALSE)
}
\arguments{
\item{times}{A list, based on the structure generated by
\code{\link{yaml_template}}}

\item{returnIntermediateTable}{Return a day-by-day summary table.}
}
\value{
If returnIntermediateTable is FALSE, a summary tibble, with three columns:
   1. deficit: Whether the total represents a deficit of total time after
               comparing to day-by-day targetDurations and the overall
               starting timelog totals.
   2. hours:   The number of hours.
   3. minutes: The number of minutes.
 If returnIntermediateTable is TRUE, a tibble, aggregated by day, with four columns:
   1. day:                    The day identifier.
   2. total_time_formatted:   The total time logged for that day.
   3. total_time_deficit:     Whether total_time_formatted is negative.
   4. target_duration:        The target duration for the day.
}
\description{
Parse times from a template of the structure generated by
\code{\link{yaml_template}}
}
\examples{
input <- yaml::read_yaml(text = "
  Total:
    deficit: no
    hours: 1
    minutes: 10

  Days:
    Day 1:
      times:
        - -2M
")
parse_times(input)

\dontrun{
input <- yaml::read_yaml(file = "time_log.yaml")
parse_times(input)
}
}
